Move useful utilities into shared testtools.py module.
