+++ flocker/acceptance/obsolete/test_containers.py
@@ -4,53 +4,22 @@
 Tests for the control service REST API.
 """
 
-import socket
-from contextlib import closing
-from json import loads
-
-from json import dumps
+from json import loads, dumps
 
 from twisted.trial.unittest import TestCase
 from twisted.python.filepath import FilePath
 from twisted.internet.defer import gatherResults
-
-from treq import get, post, content
-
-from eliot import Message
 
 from ...testtools import (
     loop_until, random_name,
 )
 from ..testtools import (
     require_cluster, require_moving_backend, create_dataset,
-    create_python_container
+    create_python_container, verify_socket, post_http_server,
+    assert_http_server,
 )
 
 CURRENT_DIRECTORY = FilePath(__file__).parent()
-
-
-def verify_socket(host, port):
-    """
-    Wait until the destionation can be connected to.
-
-    :param bytes host: Host to connect to.
-    :param int port: Port to connect to.
-
-    :return Deferred: Firing when connection is possible.
-    """
-    def can_connect():
-        with closing(socket.socket()) as s:
-            conn = s.connect_ex((host, port))
-            Message.new(
-                message_type="acceptance:verify_socket",
-                host=host,
-                port=port,
-                result=conn,
-            ).write()
-            return conn == 0
-
-    dl = loop_until(can_connect)
-    return dl
 
 
 class ContainerAPITests(TestCase):
@@ -141,8 +110,8 @@ class ContainerAPITests(TestCase):
             return d
         creating_dataset.addCallback(create_container)
         creating_dataset.addCallback(
-            lambda _: self.post_http_server(
-                node1.public_address, 8080, post_data)
+            lambda _: post_http_server(
+                self, node1.public_address, 8080, post_data)
         )
 
         def move_container(_):
@@ -152,8 +121,8 @@ class ContainerAPITests(TestCase):
             return moved
         creating_dataset.addCallback(move_container)
         creating_dataset.addCallback(
-            lambda _: self.assert_http_server(
-                node2.public_address, 8080,
+            lambda _: assert_http_server(
+                self, node2.public_address, 8080,
                 expected_response=post_data["data"])
         )
 
@@ -189,12 +158,12 @@ class ContainerAPITests(TestCase):
             return d
         creating_dataset.addCallback(create_container)
         creating_dataset.addCallback(
-            lambda _: self.post_http_server(
-                node.public_address, 8080, post_data)
+            lambda _: post_http_server(
+                self, node.public_address, 8080, post_data)
         )
         creating_dataset.addCallback(
-            lambda _: self.assert_http_server(
-                node.public_address, 8080,
+            lambda _: assert_http_server(
+                self, node.public_address, 8080,
                 expected_response=post_data["data"])
         )
         creating_dataset.addCallback(
@@ -213,8 +182,8 @@ class ContainerAPITests(TestCase):
             return d
         creating_dataset.addCallback(create_second_container)
         creating_dataset.addCallback(
-            lambda _: self.assert_http_server(
-                node.public_address, 8081,
+            lambda _: assert_http_server(
+                self, node.public_address, 8081,
                 expected_response=post_data["data"])
         )
         return creating_dataset
@@ -237,90 +206,6 @@ class ContainerAPITests(TestCase):
         creating.addCallback(created)
         return creating
 
-    def post_http_server(self, host, port, data, expected_response=b"ok"):
-        """
-        Make a POST request to an HTTP server on the given host and port
-        and assert that the response body matches the expected response.
-
-        :param bytes host: Host to connect to.
-        :param int port: Port to connect to.
-        :param bytes data: The raw request body data.
-        :param bytes expected_response: The HTTP response body expected.
-            Defaults to b"ok"
-        """
-        def make_post(host, port, data):
-            request = post(
-                "http://{host}:{port}".format(host=host, port=port),
-                data=data,
-                persistent=False
-            )
-
-            def failed(failure):
-                Message.new(message_type=u"acceptance:http_query_failed",
-                            reason=unicode(failure)).write()
-                return False
-            request.addCallbacks(content, failed)
-            return request
-        d = verify_socket(host, port)
-        d.addCallback(lambda _: loop_until(lambda: make_post(
-            host, port, data)))
-        d.addCallback(self.assertEqual, expected_response)
-        return d
-
-    def query_http_server(self, host, port, path=b""):
-        """
-        Return the response from a HTTP server.
-
-        We try multiple since it may take a little time for the HTTP
-        server to start up.
-
-        :param bytes host: Host to connect to.
-        :param int port: Port to connect to.
-        :param bytes path: Optional path and query string.
-
-        :return: ``Deferred`` that fires with the body of the response.
-        """
-        def query():
-            req = get(
-                "http://{host}:{port}{path}".format(
-                    host=host, port=port, path=path),
-                persistent=False
-            )
-
-            def failed(failure):
-                Message.new(message_type=u"acceptance:http_query_failed",
-                            reason=unicode(failure)).write()
-                return False
-            req.addCallbacks(content, failed)
-            return req
-
-        d = verify_socket(host, port)
-        d.addCallback(lambda _: loop_until(query))
-        return d
-
-    def assert_http_server(self, host, port,
-                           path=b"", expected_response=b"hi"):
-
-        """
-        Assert that a HTTP serving a response with body ``b"hi"`` is running
-        at given host and port.
-
-        This can be coupled with code that only conditionally starts up
-        the HTTP server via Flocker in order to check if that particular
-        setup succeeded.
-
-        :param bytes host: Host to connect to.
-        :param int port: Port to connect to.
-        :param bytes path: Optional path and query string.
-        :param bytes expected_response: The HTTP response body expected.
-            Defaults to b"hi"
-
-        :return: ``Deferred`` that fires when assertion has run.
-        """
-        d = self.query_http_server(host, port, path)
-        d.addCallback(self.assertEqual, expected_response)
-        return d
-
     @require_cluster(1)
     def test_non_root_container_can_access_dataset(self, cluster):
         """
@@ -342,7 +227,7 @@ class ContainerAPITests(TestCase):
         creating_dataset.addCallback(created_dataset)
 
         creating_dataset.addCallback(
-            lambda _: self.assert_http_server(node.public_address, 8080))
+            lambda _: assert_http_server(self, node.public_address, 8080))
         return creating_dataset
 
     @require_cluster(2)
@@ -381,8 +266,8 @@ class ContainerAPITests(TestCase):
             ])
 
         running.addCallback(
-            lambda _: self.assert_http_server(
-                origin.public_address, origin_port))
+            lambda _: assert_http_server(
+                self, origin.public_address, origin_port))
         return running
 
     @require_cluster(2)
@@ -409,5 +294,5 @@ class ContainerAPITests(TestCase):
 
         running.addCallback(
             # Connect to the machine where the container is NOT running:
-            lambda _: self.assert_http_server(origin.public_address, port))
+            lambda _: assert_http_server(self, origin.public_address, port))
         return running
